requirements.txt
ipaddress
dnspython
pystray